[Agent] refactor slot modal selection
